Get rid of CSDE across the code and docs
