+++ README.md
@@ -1,25 +1,22 @@
-# CSDE-cordapp-template-kotlin
+# release-V5 cordapp-template-kotlin
 
-
-To help make the process of prototyping CorDapps on Corda 5 release more straight forward we have developed the Cordapp Standard Development Environment (CSDE).
-
-The CSDE is obtained by cloning this CSDE-Cordapp-Template-Kotlin repository to your local machine. The CSDE provides:
+This template repository provides:
 
 - A pre-setup Cordapp Project which you can use as a starting point to develop your own prototypes.
 
 - A base Gradle configuration which brings in the dependencies you need to write and test a Corda 5 Cordapp.
 
-- A set of Gradle helper tasks which speed up and simplify the development and deployment process.
+- A set of Gradle helper tasks, provided by the [Corda runtime gradle plugin](https://github.com/corda/corda-runtime-os/tree/release/os/5.2/tools/corda-runtime-gradle-plugin#readme), which speed up and simplify the development and deployment process.
 
 - Debug configuration for debugging a local Corda cluster.
 
-- The MyFirstFlow code which forms the basis of this getting started documentation, this is located in package com.r3.developers.csdetemplate.flowexample
+- The MyFirstFlow code which forms the basis of this getting started documentation, this is located in package com.r3.developers.cordapptemplate.flowexample
 
-- A UTXO example in package com.r3.developers.csdetemplate.utxoexample packages
+- A UTXO example in package com.r3.developers.cordapptemplate.utxoexample packages
 
 - Ability to configure the Members of the Local Corda Network.
 
-To find out how to use the CSDE, please refer to the *Getting Started Using the CSDE* subsection within the *Developing Applications* section in the latest Corda 5 documentation at https://docs.r3.com/
+To find out how to use the template, please refer to the *<new section name goes here>* subsection within the *Developing Applications* section in the latest Corda 5 documentation at https://docs.r3.com/
 
 
 ## Chat app
@@ -31,19 +28,22 @@ In this app you can:
 3. Individually query out the history of one chat entry. `GetChatFlowArgs`
 4. Continue chatting within the chat entry with the counterparty. `UpdateChatFlow`
 
+### Prerequisites
+Corda CLI must be installed. It is used by the Corda runtime gradle plugin internally, as well as by the example Docker compose file to create necessary Kafka topics. The version of Corda CLI should match that of Corda combined worker image used in the compose file.
+
 ### Setting up
 
 1. We will begin our test deployment with clicking the `startCorda`. This task will load up the combined Corda workers in docker.
-   A successful deployment will allow you to open the REST APIs at: https://localhost:8888/api/v1/swagger#. You can test out some of the
+   A successful deployment will allow you to open the REST APIs at: https://localhost:8888/api/v5_2/swagger#. You can test out some of the
    functions to check connectivity. (GET /cpi function call should return an empty list as for now.)
-2. We will now deploy the cordapp with a click of `5-vNodeSetup` task. Upon successful deployment of the CPI, the GET /cpi function call should now return the meta data of the cpi you just upload
+2. We will now deploy the cordapp with a click of `vNodeSetup` task. Upon successful deployment of the CPI, the GET /cpi function call should now return the meta data of the cpi you just upload
 
 
 
 ### Running the chat app
 
 In Corda 5, flows will be triggered via `POST /flow/{holdingidentityshorthash}` and flow result will need to be view at `GET /flow/{holdingidentityshorthash}/{clientrequestid}`
-* holdingidentityshorthash: the id of the network participants, ie Bob, Alice, Charlie. You can view all the short hashes of the network member with another gradle task called `ListVNodes`
+* holdingidentityshorthash: the id of the network participants, ie Bob, Alice, Charlie. You can view all the short hashes of the network member with another gradle task called `listVNodes`
 * clientrequestid: the id you specify in the flow requestBody when you trigger a flow.
 
 #### Step 1: Create Chat Entry
@@ -53,7 +53,7 @@ Go to `POST /flow/{holdingidentityshorthash}`, enter the identity short hash(Ali
 ```
 {
     "clientRequestId": "create-1",
-    "flowClassName": "com.r3.developers.csdetemplate.utxoexample.workflows.CreateNewChatFlow",
+    "flowClassName": "com.r3.developers.cordapptemplate.utxoexample.workflows.CreateNewChatFlow",
     "requestBody": {
         "chatName":"Chat with Bob",
         "otherMember":"CN=Bob, OU=Test Dept, O=R3, L=London, C=GB",
@@ -70,7 +70,7 @@ Go to `POST /flow/{holdingidentityshorthash}`, enter the identity short hash(Ali
 ```
 {
     "clientRequestId": "list-1",
-    "flowClassName": "com.r3.developers.csdetemplate.utxoexample.workflows.ListChatsFlow",
+    "flowClassName": "com.r3.developers.cordapptemplate.utxoexample.workflows.ListChatsFlow",
     "requestBody": {}
 }
 ```
@@ -85,7 +85,7 @@ this message will be recorded as a message from Alice, vice versa. And the id fi
 ```
 {
     "clientRequestId": "update-1",
-    "flowClassName": "com.r3.developers.csdetemplate.utxoexample.workflows.UpdateChatFlow",
+    "flowClassName": "com.r3.developers.cordapptemplate.utxoexample.workflows.UpdateChatFlow",
     "requestBody": {
         "id":" ** fill in id **",
         "message": "How are you today?"
@@ -100,7 +100,7 @@ After a few back and forth of the messaging, you can view entire chat history by
 ```
 {
     "clientRequestId": "get-1",
-    "flowClassName": "com.r3.developers.csdetemplate.utxoexample.workflows.GetChatFlow",
+    "flowClassName": "com.r3.developers.cordapptemplate.utxoexample.workflows.GetChatFlow",
     "requestBody": {
         "id":" ** fill in id **",
         "numberOfRecords":"4"
