+++ gradle.properties
@@ -11,7 +11,7 @@ cordaNotaryPluginsVersion=5.2.0.0-RC02
 # Specify the version of the cordapp-cpb and cordapp-cpk plugins
 cordaPluginsVersion=7.0.4
 
-# Specify the version of the CSDE gradle plugin to use
+# Specify the version of the Corda runtime Gradle plugin to use
 cordaGradlePluginVersion=5.2.0.0-RC02
 
 # Specify the name of the workflows module
@@ -67,9 +67,9 @@ processorTimeout=-1
 # Value is in milliseconds
 vnodeRegistrationTimeoutDefault=30000
 
-# Specify if you want to run the contracts and workflows tests as part of the CSDE-cordapp > 3-buildCpis task
-# False by default, will execute the tests every time you stand the CSDE up - gives extra protection
-# Set to true to skip the tests, making the CSDE launching process quicker. You will be responsible for running workflow tests yourself
+# Specify if you want to run the contracts and workflows tests as part of the corda-runtime-plugin-cordapp > buildCpis task
+# False by default, will execute the tests every time you stand the template up - gives extra protection
+# Set to true to skip the tests, making the launching process quicker. You will be responsible for running workflow tests yourself
 skipTestsDuringBuildCpis=false
 
 # R3 internal repository
