+++ workflows/src/main/kotlin/com/r3/developers/cordapptemplate/flowexample/workflows/MyFirstFlow.kt
@@ -1,4 +1,4 @@
-package com.r3.developers.csdetemplate.flowexample.workflows
+package com.r3.developers.cordapptemplate.flowexample.workflows
 
 import net.corda.v5.application.flows.*
 import net.corda.v5.application.marshalling.JsonMarshallingService
@@ -146,7 +146,7 @@ class MyFirstFlowResponder: ResponderFlow {
 RequestBody for triggering the flow via REST:
 {
     "clientRequestId": "r1",
-    "flowClassName": "com.r3.developers.csdetemplate.flowexample.workflows.MyFirstFlow",
+    "flowClassName": "com.r3.developers.cordapptemplate.flowexample.workflows.MyFirstFlow",
     "requestBody": {
         "otherMember":"CN=Bob, OU=Test Dept, O=R3, L=London, C=GB"
         }
