+++ workflows/src/main/kotlin/com/r3/developers/cordapptemplate/utxoexample/workflows/GetChatFlow.kt
@@ -1,6 +1,6 @@
-package com.r3.developers.csdetemplate.utxoexample.workflows
+package com.r3.developers.cordapptemplate.utxoexample.workflows
 
-import com.r3.developers.csdetemplate.utxoexample.states.ChatState
+import com.r3.developers.cordapptemplate.utxoexample.states.ChatState
 import net.corda.v5.application.flows.ClientRequestBody
 import net.corda.v5.application.flows.ClientStartableFlow
 import net.corda.v5.application.flows.CordaInject
@@ -107,7 +107,7 @@ class GetChatFlow: ClientStartableFlow {
 RequestBody for triggering the flow via REST:
 {
     "clientRequestId": "get-1",
-    "flowClassName": "com.r3.developers.csdetemplate.utxoexample.workflows.GetChatFlow",
+    "flowClassName": "com.r3.developers.cordapptemplate.utxoexample.workflows.GetChatFlow",
     "requestBody": {
         "id":"** fill in id **",
         "numberOfRecords":"4"
